% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_flow}
\alias{data_flow}
\title{Get Dataflows for United Nations Data and Its Details}
\usage{
data_flow()
}
\description{
This function will return all dataflows (categories) in UNData database, and its several details in a dataframe.
}
\details{
The ids column represents dataflows for each variable.
}
