% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_structure}
\alias{data_structure}
\title{Get the Data Structure of a Selected Dataflow}
\usage{
data_structure(dataflow)
}
\description{
Returns a structure of the selected dataflow.
}
\details{
This function will return in a list format which includes all variables (criteria) in selected dataflow. Dataflow ids can be found in data frame that \code{data_flow()} returns.
}
\examples{
data_structure(dataflow = "DF_UNData_UNFCC")
}
