% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data}
\alias{get_data}
\title{Get Dataset for Selected Dataflow and Several Filters}
\usage{
get_data(dataflow, filter, start, end)
}
\description{
This function will return a dataframe according to the chosen \code{dataflow} and informations in \code{filter}.
}
\details{
For the filter variable one has to indicate a list which includes vectors that contains the name of variable (criteria) and selected object in this criteria.
}
\examples{
#To fetch the Environment data for Germany for the indicator of Methane (EN_ATM_METH_XLULUCF) between the years of 2010 and 2015:
\code{get_data("DF_UNData_UNFCC", filter = list(c("REF_AREA","DEU"),c("INDICATOR", "EN_ATM_METH_XLULUCF")), start = 2010, end = 2015)}
}
